# CLN: Refactor ecl scanning to use resfo