--- a/src/xtgeo/grid3d/_grid3d_utils.py
+++ b/src/xtgeo/grid3d/_grid3d_utils.py
@@ -2,14 +2,13 @@
 
 from __future__ import annotations
 
-import re
 from typing import TYPE_CHECKING, Literal
 
 import numpy as np
 import pandas as pd
+import resfo
 import roffio
 
-from xtgeo import XTGeoCLibError, _cxtgeo
 from xtgeo.common import null_logger
 from xtgeo.common.constants import MAXDATES, MAXKEYWORDS
 
@@ -36,24 +35,13 @@ def scan_keywords(
     if fformat not in ("xecl", "roff"):
         raise ValueError(f"File format can be either `roff` or `xecl`, given {fformat}")
 
-    if fformat == "xecl":
-        pfile.get_cfhandle()  # just to keep cfhanclecounter correct
-        if dates:
-            keywords = _scan_ecl_keywords_w_dates(
-                pfile,
-                maxkeys=maxkeys,
-                dataframe=dataframe,
-            )
-        else:
-            keywords = _scan_ecl_keywords(
-                pfile,
-                maxkeys=maxkeys,
-                dataframe=dataframe,
-            )
-        pfile.cfclose()
-    elif fformat == "roff":
-        keywords = _scan_roff_keywords(pfile, maxkeys=maxkeys, dataframe=dataframe)
-    return keywords
+    if fformat == "roff":
+        return _scan_roff_keywords(pfile, maxkeys=maxkeys, dataframe=dataframe)
+    return (
+        _scan_ecl_keywords_w_dates(pfile, maxkeys=maxkeys, dataframe=dataframe)
+        if dates
+        else _scan_ecl_keywords(pfile, maxkeys=maxkeys, dataframe=dataframe)
+    )
 
 
 def scan_dates(
@@ -65,37 +53,28 @@ def scan_dates(
 
     Cf. grid_properties.py description
     """
+    dates = []
+    seqnum = -1
+    for item in resfo.lazy_read(pfile.file):
+        kw = item.read_keyword().strip()
+        data = item.read_array()
 
-    seq = _cxtgeo.new_intarray(maxdates)
-    day = _cxtgeo.new_intarray(maxdates)
-    mon = _cxtgeo.new_intarray(maxdates)
-    yer = _cxtgeo.new_intarray(maxdates)
+        if kw == "SEQNUM":
+            seqnum = data[0]
+            continue
 
-    cfhandle = pfile.get_cfhandle()
-
-    nstat = _cxtgeo.grd3d_ecl_tsteps(cfhandle, seq, day, mon, yer, maxdates)
-
-    pfile.cfclose()
-
-    sq = []
-    da = []
-    for i in range(nstat):
-        sq.append(_cxtgeo.intarray_getitem(seq, i))
-        dday = _cxtgeo.intarray_getitem(day, i)
-        dmon = _cxtgeo.intarray_getitem(mon, i)
-        dyer = _cxtgeo.intarray_getitem(yer, i)
-        date = f"{dyer:4}{dmon:02}{dday:02}"
-        da.append(int(date))
-
-    for item in [seq, day, mon, yer]:
-        _cxtgeo.delete_intarray(item)
-
-    zdates = list(zip(sq, da))  # list for PY3
+        # With LGRs multiple INTEHEADs may occur. Ensure we get the date
+        # from the first INTEHEAD after a SEQNUM.
+        if kw == "INTEHEAD" and seqnum != -1:
+            # Index 66 = year, 65 = month, 64 = day
+            date = int(f"{data[66]}{data[65]:02d}{data[64]:02d}")
+            dates.append((seqnum, date))
+            seqnum = -1
 
     return (
-        pd.DataFrame.from_records(zdates, columns=["SEQNUM", "DATE"])
+        pd.DataFrame.from_records(dates, columns=["SEQNUM", "DATE"])
         if dataframe
-        else zdates
+        else dates
     )
 
 
@@ -104,52 +83,25 @@ def _scan_ecl_keywords(
     maxkeys: int = MAXKEYWORDS,
     dataframe: bool = False,
 ) -> list[KeywordTuple] | pd.DataFrame:
-    cfhandle = pfile.get_cfhandle()
-
-    # maxkeys*10 is used for 1D keywords; 10 => max 8 letters in eclipse +
-    # "|" + "extra buffer"
-    nkeys, keywords, rectypes, reclens, recstarts = _cxtgeo.grd3d_scan_eclbinary(
-        cfhandle, maxkeys * 10, maxkeys, maxkeys, maxkeys
-    )
-
-    pfile.cfclose()
-
-    if nkeys == -1:
-        raise XTGeoCLibError(f"scanning ecl keywords exited with error code {nkeys}")
-
-    if nkeys == -2:
-        raise XTGeoCLibError(
-            f"Number of keywords seems greater than {maxkeys} which is currently a "
-            "hard limit"
+    keywords = []
+    for item in resfo.lazy_read(pfile.file):
+        keywords.append(
+            (
+                item.read_keyword().strip(),
+                item.read_type().strip(),
+                item.read_length(),
+                item.stream.tell(),
+            )
         )
 
-    keywords = re.sub(r"\s+\|", "|", keywords)
-    keywords = keywords.split("|")
-
-    # record types translation (cf: grd3d_scan_eclbinary.c in cxtgeo)
-    rct = {
-        1: "INTE",
-        2: "REAL",
-        3: "DOUB",
-        4: "CHAR",
-        5: "LOGI",
-        6: "MESS",
-        -1: "????",
-    }
-
-    rectypes = rectypes.tolist()
-    rc = [rct[key] for nn, key in enumerate(rectypes) if nn < nkeys]
-    rl = reclens[0:nkeys].tolist()
-    rs = recstarts[0:nkeys].tolist()
-
-    result = list(zip(keywords, rc, rl, rs))
-
-    if dataframe:
-        cols = ["KEYWORD", "TYPE", "NITEMS", "BYTESTART"]
-        df = pd.DataFrame.from_records(result, columns=cols)
-        return df
-
-    return result
+    return (
+        pd.DataFrame.from_records(
+            keywords,
+            columns=["KEYWORD", "TYPE", "NITEMS", "BYTESTART"],
+        )
+        if dataframe
+        else keywords
+    )
 
 
 def _scan_ecl_keywords_w_dates(
@@ -176,12 +128,14 @@ def _scan_ecl_keywords_w_dates(
         entry = (name, dtype, reclen, bytepos, date)
         result.append(entry)
 
-    if dataframe:
-        cols = ["KEYWORD", "TYPE", "NITEMS", "BYTESTART", "DATE"]
-        df = pd.DataFrame.from_records(result, columns=cols)
-        return df
-
-    return result
+    return (
+        pd.DataFrame.from_records(
+            result,
+            columns=["KEYWORD", "TYPE", "NITEMS", "BYTESTART", "DATE"],
+        )
+        if dataframe
+        else result
+    )
 
 
 def _scan_roff_keywords(
@@ -235,11 +189,14 @@ def _scan_roff_keywords(
             if done:
                 break
 
-    if dataframe:
-        cols = ["KEYWORD", "TYPE", "NITEMS", "BYTESTARTDATA"]
-        return pd.DataFrame.from_records(keywords, columns=cols)
-
-    return keywords
+    return (
+        pd.DataFrame.from_records(
+            keywords,
+            columns=["KEYWORD", "TYPE", "NITEMS", "BYTESTARTDATA"],
+        )
+        if dataframe
+        else keywords
+    )
 
 
 def _get_roff_type_and_size(