Resolves #838 

Some benchmarks

```sh
# xtgeo-testdata/3dgrids/simpleb8/E100_3LETTER_TRACER.UNRST - 1.01 MB
OLD: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.0027261669747531414 seconds
NEW: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.002314541954547167 seconds

# xtgeo-testdata/3dgrids/reek/REEK.UNRST - 42.4 MB
OLD: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.015472916886210442 seconds
NEW: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.06493925000540912 seconds

# xtgeo-testdata/3dgrids/bench_spe9/BENCH_SPE9.UNRST - 90.7 MB
OLD: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.0327368329744786 seconds
NEW: tests/test_grid3d/test_grid_properties.py::test_scan_dates Dates scanned in 0.055844167014583945 seconds
```

This is a limited data set but it looks like resfo may not slow down as quickly as the C code relative to the file size.